mas arreglos de arrays

+++ src/controlstock/BebidasPerdidas.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 
 public class BebidasPerdidas {
 
-    public static void bebidasPerdidas(double[] Sbebida, double valorB[], String[] bebida) {
+    public static void bebidasPerdidas(double[] Sbebida, int valorB[], String[] bebida) {
         int i, opcion, aux;
         Scanner teclado = new Scanner(System.in);
         System.out.println("");
