+++ src/controlstock/MenuSnack.java
@@ -6,7 +6,7 @@ package controlstock;
 
 public class MenuSnack {
 
-    public static void menuSnack(double[] Ssnack, double valorSNK[], String[] Snack) {
+    public static void menuSnack(double[] Ssnack, int valorSNK[], String[] Snack) {
         int i;
         // Mostramos todo el menú de snacks y la cantidad de stock disponible.
         System.out.println("---------------------------------------");
